Feat : Crud 구현

+++ src/main/java/com/example/yamyam16/store/dto/request/UpdateStoreRequestDto.java
@@ -13,7 +13,7 @@ public class UpdateStoreRequestDto {
     private String storename;
     private Long open_time;
     private Long close_time;
-    private int minOrderPrice;
+    private Long minOrderPrice;
     private String category;
     @Size(max = 100, message = "공지는 30글자 이하로 입력해주세요.")
     private String notice;
